fix(server): authValues needs to be preserved in client instead of in envelope.
envelope.Values would be reset in ResetTransaction(). If the client is reused (means envelope.Values would be reset after DATA command),
the next transaction would get an empty envelope.Values which is unexpected.

We keep authValues in client; copy it to envelope.Values in handleData() so processors could still access authValues.